Updated version with face recognition
This version is able to recognize faces using the relational size and distance between features. Saved faces are recorded in "persons.json" file as person objects. When the application is launched, all profiles detected in the image will be compared to the faces (person objects) in the json file, and if they are similar enough the display will show their name and how many features overlapped. At the moment there is no way to create or modify the json file; that's something that should probably be added. This version also does not include the color detection that's currently in the main branch, so that could be implemented as well.